Include `onchain gas` cost (#1195)
<!-- Reference any GitHub issues resolved by this PR -->

Closes #1186 

## Introduced changes

<!-- A brief description of the changes -->

- onchain data gas is computed from the state change from blockifier
state

## Checklist

<!-- Make sure all of these are complete -->

- [x] Linked relevant issue
- [x] Updated relevant documentation
- [x] Added relevant tests
- [x] Performed self-review of the code
- [x] Added changes to `CHANGELOG.md`

---------

+++ crates/forge-runner/src/lib.rs
@@ -333,8 +333,8 @@ impl TestCaseSummary<Fuzzing> {
                         _ => unreachable!(),
                     });
 
-                let max = gas_usages_vec.clone().reduce(f64::max).unwrap();
-                let min = gas_usages_vec.reduce(f64::min).unwrap();
+                let max = gas_usages_vec.clone().reduce(u128::max).unwrap();
+                let min = gas_usages_vec.reduce(u128::min).unwrap();
 
                 TestCaseSummary::Passed {
                     name,
