+++ crates/forge-runner/src/test_case_summary.rs
@@ -9,8 +9,8 @@ use test_collector::{ExpectedPanicValue, ExpectedTestResult};
 
 #[derive(Debug, PartialEq, Clone)]
 pub struct GasStatistics {
-    pub min: f64,
-    pub max: f64,
+    pub min: u128,
+    pub max: u128,
 }
 #[derive(Debug, PartialEq, Clone)]
 pub struct FuzzingStatistics {
@@ -32,7 +32,7 @@ impl TestType for Fuzzing {
 #[derive(Debug, PartialEq, Clone)]
 pub struct Single;
 impl TestType for Single {
-    type GasInfo = f64;
+    type GasInfo = u128;
     type TestStatistics = ();
 }
 
@@ -150,7 +150,7 @@ impl TestCaseSummary<Single> {
         test_case: &TestCaseRunnable,
         arguments: Vec<Felt252>,
         fork_info: &ForkInfo,
-        gas: f64,
+        gas: u128,
     ) -> Self {
         let name = test_case.name.to_string();
         let msg = extract_result_data(&run_result, &test_case.expected_result);
